requirements.txt
fastapi
uvicorn
feedparser
google-generativeai
python-dotenv
beautifulsoup4
requests
openai
google-genai
